pyproject.toml
[project]
name = "token-guard"
version = "0.1.0"
description = ""
authors = [
    { name = "Chathura Matheesha", email = "chathuramatheesha@gmail.com" }
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi (>=0.115.12,<0.116.0)",
    "uvicorn (>=0.34.2,<0.35.0)",
    "sqlalchemy[asyncio] (>=2.0.41,<3.0.0)",
    "aiosqlite (>=0.21.0,<0.22.0)",
    "python-jose (>=3.4.0,<4.0.0)",
    "argon2-cffi (>=23.1.0,<24.0.0)",
    "pydantic[email] (>=2.11.5,<3.0.0)",
    "pydantic-settings (>=2.9.1,<3.0.0)",
    "ulid-py (>=1.1.0,<2.0.0)"
]

[tool.poetry]
packages = [{ include = "token_guard", from = "src" }]


[tool.poetry.group.dev.dependencies]
black = "^25.1.0"

[tool.poetry.scripts]
start = 'token_guard.cli:start'
dev = 'token_guard.cli:dev'
test = 'pytest:main'

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
